net: shell: dns: Properly manage lifetime of callback data
DNS callback needs "struct shell *shell" data structure to pass as
a parameter to shell print. The obvious way to achieve that would be
to store in in stack structure, but turns out we allow system shell
to be run in more than one concurrent instances.

Then it should be kept on stack as before, but we should block in
the main handler until the callback returns. Alas, that doesn't
work either, shell print routines deadlock in this case.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>