drivers: timer: nrf: Use non-blocking LF clock start
As clock_control_on is blocking, using it during startup would
lead to hundred of milliseconds waiting time. Switched to use
non blocking start to allow other drivers initialization which
is performed in the same context.

Moreover, blocking call is using semaphore and it's not allowed
in pre_kernel phase.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>