drivers: sensor_monitor: Add periodical monitor of sensor channel
Added a module which implements a use case where user need to monitor
a sensor channel (e.g. temperature) periodically to perform an action
if sensor value changes (or not). User provides a decision function
which is called on every measurement. Function decides if action shall
be taken. Action is user defined. Additionally, user can define maximum
number of action skips after which action is triggered.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>