logging: Use runtime packaging in immediate mode
Removed Z_LOG_MSG2_SYNC macro which was attempting to create
log message (and cbprintf package) on stack and the process
it in this context (immediate mode). Since processing happens
in that context there is no significant gain when using
static packaging. Additionally, this approach does not work
in userspace.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>