wip: modify hello_world to validate that system timer behavior
Due to reports that there is an issue with system timer when uptime
exceeds 32bits this sample is modified. It is using nrf_rtc_timer
feature to shift time further in future by ~0xffffff. It is done
multiple times to go beyond 32 bit time and then k_sleep is
validated. Counter based on RTC2 is used to check that thread
wakes up correctly.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>