samples: Add ppr_stress sample which crashes on ppr
Sample is using 2 priority levels: lower for GRTC and higher
for counter (TIMER) instance. Sample is randomly starting kernel
and counter timeouts and it crashes almost immediately.

Sample can be build for cpuapp and it works fine. If priority is
adjusted in the overlay file so that GRTC and TIMER use the same
priority then it also works."

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>